[BUGFIX] Ensure package state file using the setup command
With #99221 the cli setup command has been introduced to
setup a instance from the command line and improved with
followup changes.

This works pretty fine in composer mode installations. In
non-composer mode installations the package state file is
not created using the cli command whereas the web based
installation process creates one.

This change ensures that a package state file is created
if none exists. If no file is found, it creates it based
on default factory settings like the installer.

Resolves: #101678
Related: #100719
Related: #99221
Releases: main, 12.4
Change-Id: I022c4300131d0eca82569ef0463d0c0078c53055
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80900
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>