[TASK] Suggest PHP exif module in system environment status
If the PHP exif module is installed, TYPO3 uses the `exif_read_data()`
function of read exif metadata of images. In case of the image upload
functionality in TYPO3, the `exif_read_data()` function is used to
evaluate the `Orientation` exif property, which is used to decide, if
the width and height of an image should be switched before the data
is saved in the `sys_file_metadata` table. If the PHP exif module is
not available, portrait images created by modern camera systems might
be saved with wrong width and height, because the exif `Orientation`
is not respected. This might result in rending of distorted images,
when the uploaded image is processed by TYPO3 image rendering.

With this change, the system environment check is extended to
suggest to PHP exif module to be installed. If not available,
a warning message is shown in the system environment check.
Additionally, composer.json files of system extensions using
exif related functions (is available) have been updated to
suggest PHP ext-exif.

Resolves: #101922
Releases: main
Signed-off-by: Torben Hansen <derhansen@gmail.com>
Change-Id: I26400c47b66185a0236bf72b9010d153c328c70b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81022
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Guido Schmechel <guido.schmechel@brandung.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Guido Schmechel <guido.schmechel@brandung.de>