[BUGFIX] Preserve native CKEditor5 related toolbar settings
If the CKEditor is configured with the new v5-style
`toolbar.removeItems` option instead of `removeButtons` (v4), the
CKEditor5Migrator will now merge`removeItems` with `removeButtons`.

Also do not overwrite `toolbar.shouldNotGroupWhenFull`, allowing
it to be set to false.

Note that v5-style `toolbar.items` where already defined to
be used when v4-style `toolbarGroups` are empty, so we do
not change this behavior.

Resolves: #100829
Resolves: #101869
Releases: main, 12.4
Change-Id: Iebf70567866af13411c3d3b734e2f518dbd0b4e3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78959
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>