[TASK] Replace icon for password reset in ext:beuser
The current icon on the password reset button in ext:beuser does
not clearly visualize the functionality behind the button. Besides
that, the same icon is used in several other places in TYPO3 on
buttons, which most likely have either undo or revert functionality.

This patch replaces the current icon with the `actions-key` icon,
which might raise more awareness to the user, that the functionality
behind the button has something to do with the user password.

Resolves: #101932
Releases: main
Signed-off-by: Torben Hansen <derhansen@gmail.com>
Change-Id: I7a42c1592736825f5ebf541ee9b67a6357cd4d12
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81032
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>