[TASK] Move GraphicalFunctions code to ImageProcessingInstructions
This change moves some logic to calculate width
and height based on instructions from an incoming
array into a separate DTO, which can be used consistently
and type-safe instead of residing in GraphicalFunctions.

Some transformation of options ("maxW" vs. "maxWidth")
is now moved to this new class as well, in order to
centralize the naming of the options.

In addition, the special calculation for ImageMagick
about the cropH and cropV values is now done directly
in the instructions, so the CropArea object can now
be used in more places than before.

Resolves: #101957
Releases: main
Change-Id: I14608f486e4e3bf5cf669528f01d03479b8e6e93
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81050
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>