[TASK] Streamline applying stylesheets in lit-element components
This change extracts dynamic stylesheets of web components that
are attached to the light DOM to a corresponding static file in
the directory `Build/Sources/Sass/element/` (it will be merged
to `backend.css` during the build process).

This change allows to get rid of the `window.litNonce` helper
function that was introduced to allow CSP for dynamic styles.

Besides that, the shadow DOM component `typo3-backend-spinner`
now uses constructed stylesheets directly. In addition, the
missing component names have been added to the global map.

Resolves: #101744
Releases: main, 12.4
Change-Id: I7cf3cdc31d0ffe685897e48f000f18e9b5b5aca6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80664
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>