[BUGFIX] Always fetch necessary fields for "record saved" notification
Due to another couple of use cases, the necessary
fields to generate the record title, used in the
"record saved" notification in FormEngine, are now
always fetched, if not already present.

Additionally, it's now ensured that the "uid" is
always available, since it is necessary for
labels with foreign table lookups.

Resolves: #101964
Releases: main, 12.4
Change-Id: I933211f582e31c7af9fea909075420316a81a544
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81095
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>