[BUGFIX] Handle missing image dimensions gracefully
Just like SvgImageProcessor, DeferredBackendImageProcessor now also
catches the exception, which is thrown, when image dimensions equal
zero and falls back to 64x64 image dimension. This enables editors to
correct bad image cropping without the need to do it in the database.

To be more specific, a new Exception ZeroImageDimensionException is
added, which is used in both processors now.

The image cropping editor is already fixed to only allow a minimum
dimension of 10x10. However, already corrupted files can not be
accessed without this patch.

Resolves: #93807
Related: #99943
Releases: main, 12.4, 11.5
Change-Id: I259c365052dd02110d43418ba48b768bbe073fc1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81111
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>