[BUGFIX] Use "sample" instead of "useSample" in FAL processing
When using TypoScript, the setting "sample" allows
to scale faster than with "geometry" setting.

However, when FAL was introduced the setting was called
"useSample" which never worked since TYPO3 v6.0 in
conjunction with FAL.

The option is however useful and is now working again.

The according code is moved into GraphicalFunctions
in order to avoid future access to this public property.

Resolves: #101981
Releases: main, 12.4
Change-Id: I5c926460d4c5393c418d1e90fcb11052ab3e5cae
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81115
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>