[DOCS] Add PHP warning impact for usergroup_cached_list field
The change log previously only mentioned the database impact.
But code might access the field as array key on the user array.
This would result in unexpected behaviour.
Furthermore this would result in Warnings since PHP 8.0.

The changelog is extended with the impact.
That allows people to not only check database queries but also array
access while migrating.

Furthermore suggested migration class was replaced by the FQCN to ease
migration. People do no longer need to find the FQCN. Furthermore it
might exist a class with same name in the project from third party code,
or it might be added another class in the future.

Resolves: #101987
Releases: main, 12.4, 11.5
Change-Id: I475606e0f9a325880ee10339adaa2971985b05fa
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81124
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>