[BUGFIX] Ensure scaled width and height are used with `noScale=1`
If imgResource option `noScale = 1` [1] is configured,
the image itself is not scaled. However, the calculated
scaled width and height values needs to be used for the
`img` tag attributes `width` and `height`.

In that case, the orginal image is used instead of a
processed image, and the calculated values are discarded.

This change recalculates these values if `noScale = 1` is
requested to ensure that the scaled width and height values
are used correctly.

Example
-------

page.100 = IMAGE
page.100 {
  file = fileadmin/styleguide/bus_lane.jpg
  file.noScale = 1
  file.width = 240m
  file.height = 240m
  wrap = |<br>
}

[1] https://docs.typo3.org/m/typo3/reference-typoscript/main/en-us/Functions/Imgresource.html#noscale

Resolves: #100972
Releases: main, 12.4
Change-Id: Ib09e3e837225347ee2e2a9c9979b014d8fd2b850
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81146
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>