[BUGFIX] Enable "hide_in_menu" toggle in ContextMenu for custom doktypes
Since TYPO3 v10 it is possible to use any number of
doktype with "pages.doktype" and the limitation of custom
doktypes "less than 200" was lifted.

One part is still missing - the ContextMenu still
has the limitation of only showing up for doktypes
less than 200.

This patch changes this behavior.

Resolves: #101999
Releases: main, 12.4
Change-Id: I6971c9856297bc669d32cdf778be1883a1be54ed
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81147
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>