[BUGFIX] Fix performance issue when resolving Extbase relations in workspaces
In the workspace preview, when Extbase resolves a MM relation
(e.g. categories from a news record), the database query introduced by
commit a1f22a427f3bc49ab3fc3424859a7e2d7beaa1c9 forgets to constrain the
MM on the foreign side of the relation.

This can result in a lot of rows being fetched from the database, which
can be a huge performance issue depending on the number of relations in
the MM table.

This change enforces the foreign clause on the foreign side of the
relation.

Resolves: #99244
Related: #88021
Related: #82750
Related: #82086
Related: #76993
Related: #57169
Releases: main, 13.4, 12.4
Change-Id: I12b354c37d1e430d201afb4e08ecb2e2acf1f6ba
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87626
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>