[BUGFIX] Show but preserve invalid classes in link-browser
In case a link with an invalid CSS class is edited,
the class value is not simply invalidated and shown
as empty (which isn't true), but a virtual select option
"[ INVALID VALUE (…) ]" is added (and selected) that
documents the fact that the current value invalid,
but will be preserved upon updates. (as in FormEngine)

This allows the editor to choose between preserving,
removing or changing to another css class value.

With this change applied, we can not properly
remove empty value. Note that empty values previously
implicitly preserved existing values, which made
it impossible to remove a link title or target.

Resolves: #102012
Resolves: #102013
Releases: main, 12.4
Change-Id: Iefae377a9e18467e8eb9111ef3ef7a96c89b6ce1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81176
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>