[DOCS] Add AfterUserLoggedInEvent to list of events in felogin docs
With the change in #101933 the AfterUserLoggedInEvent is also dispatched
when a frontend user is successfully logged in. Therefore, this event
is added to the list of available events (although that event is not
in the felogin namespace) to ease the access for the developer who looks
for appropriate events regarding felogin.

Resolves: #102030
Related: #101933
Releases: main
Change-Id: I26bcc1cbc7648e91eeef9b9bd2564afcb8ebc268
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81216
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Oliver Bartsch <bo@cedev.de>