[FEATURE] Allow redirect filtering by "protected" state
The redirects module does currently not allow to filter redirects
based on the `protected` state. This making it hard to effectively
administrate protected redirects on websites with a huge amount of
redirects.

This change introduces a new filter option in the redirect
administration module which makes it possible to filter redirects
based on the `protected` state.

Protected redirects are now also visually highlighted by a lock-icon
in the list of redirects.

Additionally, unit- and functional-tests have been extended to
cover redirect filtering based on the `creation_type` and
`protected` state.

Resolves: #102072
Releases: main
Signed-off-by: Torben Hansen <derhansen@gmail.com>
Change-Id: I4537c2a4ab09e78ac72cc63a393fa28558180e26
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81303
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>