[BUGFIX] Enable configuration passthrough for custom CKEditor5 plugins
Use JavaScript object destructuring to "remove" all options from the
configuration object, that are consumed by our CKEditor5 wrapper.
Also cleanup RTE resource and config handling to not write
unused options into the configuration array and to avoid resolving
legacy (CKEditor4) resource paths which are dropped anyway.

As a drive-by, dots are now substituted with '_' in RTE field IDs to
avoid the following hassle:
Dots are interpreted as CSS classes when the ID value is used
in combination with a number sign (#) to create a CSS selector
for the respective field ID. That means the selector will not match.
The class selector additionally becomes invalid once there is a digit
after the dot, as CSS classes need to start with strings.
(Example: EXT:styleguide in_flex » tab » rte.2).

Resolves: #100784
Resolves: #101437
Releases: main, 12.4
Change-Id: I7f541a39120018027792a2ad978935f743b9b6c1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81381
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>