[TASK] Improve ReferenceIndex details
The ReferenceIndex class with table 'sys_refindex' plays
a more and more important role in core when it comes to
systematic performance improvements.

We've had quite some work on sys_refindex in v11 already,
but the main class has not been touched much, yet.

It's time to untangle some more to unblock further works.

The patch brings some general changes to the @internal
marked class ReferenceIndex:

* Rely on DI
* Set PHP types that can be forced easily
* Tame more class state
* Streamline the CLI main method used with
  > bin/typo3 referenceindex:update

The changes in updateIndex() improve the overall output
of the CLI processing: For instance, tables are now
processed alphabetically, and there are changes like
more effective calls for tables that have no relations.

Resolves: #102141
Releases: main
Change-Id: Ied84f0ab6deab65592ebc1e694f1797d81c138f7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81408
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>