[BUGFIX] Allow WS preview links to preview hidden pages
When a hidden page is previewed in the Backend,
the PreviewSimulator has a valid BE_USER object
in place. However, when a preview link is generated,
the PreviewUser does not have a group, thus the PreviewSimulator
is not activated.

For this reason, the check if the PreviewSimulator
is active, needs to consider if we are in an offline workspace

Resolves: #102216
Releases: main, 12.4
Change-Id: I95aadaad1de69ce6eed91acd82f4d42171a6500d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81499
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Andreas Wolf <andreas.wolf@typo3.org>
Reviewed-by: Andreas Wolf <andreas.wolf@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>