[TASK] Simplify ImageProcessingInstruction object
The introduction of the ImageProcessing object
made it possible to refactor and clean up further
code within TYPO3's Image processing lowlevel code.

This change removes some previously added
arguments and options such as the file extension,
which needs to go into a separate resolver functionality.

In the same effort, the logic and usage of the infamous
$info array is reduced, and a drive-by fix for adding
the quality parameter multiple times (happened in
one of the previous refactorings) is added.

Resolves: #102228
Releases: main
Change-Id: Ifd7855236e73a225af402a359292f1256d591a5d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81507
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>