[TASK] Pin to PER Coding Style v1.0
PER-CS2.0 was added as clone of PER-CS1.0 in php-cs-fixer v3.27
in order to implement the new rules that are defined in
PER Coding Style 2.0 [1]. The alias PER already points to PER-CS2.0
although the current set of rules are not implemented yet. [2]
There is only one rule (single_line_empty_body) that is done by now
and more will follow, so we rather pin to PER-CS1.0 for now (and
enable single_line_empty_body which we already migrated to) in
order to prevent having coding style changes with every
php-cs-fixer raise. We rather wait until the 2.0 preset has been
finished and is on par with the upstream php-fig definitions and
then switch with one change.

[1] https://www.php-fig.org/per/coding-style/
[2] https://github.com/PHP-CS-Fixer/PHP-CS-Fixer/issues/7247

Resolves: #102240
Related: #102236
Releases: main, 12.4, 11.5
Change-Id: I33377e4e5016f1e5cc8ab08a9e71a933dcd41d93
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81534
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>