[TASK] Avoid calling migrateFlexFormTcaRecursive() twice
With #99756, migrating DS in FlexForms has (indirectly)
been added to parseDataStructureByIdentifier(), so any
DS is "clean" afterwards. FormEngine related ajax
controllers trigger this when a data structure
identifier is supplied.

Calling this a second time in FormEngine TcaFlexPrepare
does not make sense and is removed now.

This allows us to protect that internal method.

Resolves: #102265
Related: #99756
Releases: main
Change-Id: I4097cf350457212a4551634cc10d314091f3fe7d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81590
Tested-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>