[!!!][TASK] Streamline File Processing API
This change the concept of the Processed File API,
which was introduced back in TYPO3 v6 in 2012
(commit #66c204f0bf3e83d302e6a5f9e367ebda267a015e),
and was almost not touched since then.

The "TaskInterface" was hardly extended for purposes
other than "ImagePreview" (BE) and "ImageCropScaleMask" (FE).

In order to accommodate for what is / was useful
in the past years, the TaskInterface's interface
signature was adapted, next to typed methods and arguments,
the method "sanitizeConfiguration" is added,
allowing to reduce logic in the "LocalPreviewHelper".

The AbstractGraphicalTask was removed as it
extended from AbstractTask and had little value
over the concrete implementations, and wasn't even
abstract. The AbstractTask now has two abstract methods
"getName" and "getType" in order to be extended in its
subclasses.

All in all, the main goal for this patch is
to make the structure of this functionality easier
to understand:
* A task holds the configuration of a file
* A ProcessedFile object is the instance of a file that was or is about to be processed file
* A processor can be used to do the actual action

Resolves: #102165
Releases: main
Change-Id: Ica3ffcfc9c2fee5b5344f5e8b968989fd4371eb4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81618
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>