[TASK] Simplify check for ResourceCompressor HTML5 doctype
Even though ResourceCompressor is not used
in BE context anymore, the check for HTML5 doctype
is now based on PageRenderer instead of TSFE
as it is available as a singleton and not
as a global dependency.

In addition, this reduces a cross-dependency
between EXT:frontend and EXT:core.

Introducing DI etc. for ResourceCompressor
can follow at a later stage.

Resolves: #102297
Releases: main
Change-Id: I45c73e59456552e35bfa700990124d45389edb09
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81636
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>