[FEATURE] Allow custom translations for extbase validators
Extbase validators use hardcoded translation keys to display messages
when a validated value is not valid. It is however only possible to
overwrite the translation keys globally, which in many scenarios is
not suitable. It is also not possible to define an individual
translation key on property or validator basis, so translated
messages can not be individualized for the current usage scope.

This feature introduces the possibility to define custom translation
keys or a custom message for extbase validators by using validator
options. This allows developers to provide individual translation
keys for the current usage scope as shown in the example below:

```
    'validator' => 'NotEmpty',
    'options' => [
        'nullMessage' => 'LLL:EXT:my_extension/Resources/Private/Language/locallang.xlf:validation.myProperty.notNull',
        'emptyMessage' => 'LLL:EXT:my_extension/Resources/Private/Language/locallang.xlf:validation.myProperty.notEmpty',
    ],
])]
protected string $title = '';
```

This change also deprecates the validator option `errorMessage`
of the `RegularExpressionValidator` in favor to the new
`message` option.

Additionally, 3 unused translation keys in the locallang.xlf
file have been removed.

Resolves: #102326
Releases: main
Signed-off-by: Torben Hansen <derhansen@gmail.com>
Change-Id: Ie1932c7d51a29ac5655adf2ce226b2bd44a682f4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81696
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>