[BUGFIX] Fixes array_key_exists(): Argument #1 ($key) must be a valid array offset type
Fixes the error
array_key_exists(): Argument #1 ($key) must be a valid array offset type

This error occurs, for example, when the SaveToDatabase finisher tries to map a field with the name "type". In transformMultiValuePropertiesForFormEditor there is no check as to whether the form element type determined there is actually an identifier (string) or not.