[BUGFIX] Fix PHP error in FormEditor property preparation
This change fixes the error

`array_key_exists(): Argument #1 ($key) must be a valid array offset
type`

This error occurs, for example, when a SaveToDatabase finisher is
configured to map a form field named "type":

  finishers:
    -
      identifier: SaveToDatabase
      options:
        -
          table: tx_mytable
          elements:
            type:
              mapOnDatabaseColumn: foo

The FormEditor invokes
transformMultiValuePropertiesForFormEditor($definition, 'type', $types)
with the intent to transform all structures that contain a
`type: {someDefinedType}` child-entry to a slightly different format.
It does this recursively for the entire form configuration, and
failed for the finisher cases mentioned above.

As `type` has an array value in this case:

  ['mapOnDatabaseColumn' => 'foo']

it can not be used as `$key` argument for `array_key_exists`.

Releases: main, 12.4
Resolves: #102340
Change-Id: If79a96b7c22a4a39d211d3e09d11381db3de1134
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81776
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>