[BUGFIX] Create symlinks for public assets with specific vendor names
When `composer install|dumpautoload` is performed, the
PackageArtifactBuilder inspects the list of all extensions, and
then creates symlinks to `Resources/Public` of each extension
into the `_assets` subdirectory.

This step could fail, when a vendor name ended on "ext".
A package in `vendor/superagency-next/myextension/Resources/Public`
would not have gotten symlinked, because "ext/myextension"
matches in this combination.

The fix is to check the directory for the actual pattern
of non-composer legacy installations (which is available as a constant).

Resolves: #101590
Releases: main, 12.4
Change-Id: Ie19dae86dc69ce9c78681e8011575f12a46c5f8f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81782
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>