[TASK] Fix changelog for TCA slug generatorOptions
TCA fields of type 'slug' can be configured with the property
generatorOptions.

Particularly, the short form of using a simple array for
the fields instead of a nested array is a common source
of confusion and the initially introduced changelog had
incorrect examples.

The changelog is improved as follows

- Fix the examples. The examples 2 and 3 showed incorrect
  results
- Use the multidimensional array in the source snippet
- Clarify that ['nav_title', 'title'] is the same as
  [['nav_title'], ['title']], but not the same as
  [['nav_title', 'title']]

Resolves: #89390
Releases: main, 12.4, 11.5
Change-Id: Ib1944c2a1eb9e2a69a0f169f316208cf46f9be84
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81726
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>