[BUGFIX] Keep ValuePicker in sync with field value for replace mode
A "valuePicker" TCA configuration allows to either prepend, append or
replace a preset for an input field.

While "prepend" and "append" allow to continuously pick and choose
elements to be inserted, the "replace" mode is single-use and always
resets to a single chosen value.

For the "replace" mode it could be helpful to keep the input field
in sync with the ValuePicker especially if the value picker as a
human-readable representation of the value that is added in the
input field.

Resolves: #102330
Releases: main, 12.4
Change-Id: I07a8c8ca29877f2a1f580f44ed572280e87dded4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81742
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>