[BUGFIX] Ensure title is set for ElementBrowser
FieldControl is required to set a title.
To ensure this, the field control options
can now be used to set a custom title.

Additionally, a fallback is added to ensure
a title is always set, even if a custom
TCA type is used and no title is defined,
using the FieldControl option.

Resolves: #102366
Releases: main, 12.4
Change-Id: Id4b2bbb15d5e1d12fae6dbd287dbdeda7ba5b987
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81746
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Jochen Roth <rothjochen@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen Roth <rothjochen@gmail.com>