[BUGFIX] Correctly re-import extension static-data
The ext_tables_static+adt.sql data are now re-imported,
if its data changed, when setting up the extensions.

This happens either from within the Extension Manager,
or via the CLI  command.

Moreover, a legacy value, which was already taken care of
in `ReloadSqlDataViewHelper`, is now respected as well.

Resolves: #102376
Releases: main, 12.4
Change-Id: I67074f8b98acd0bbfc7ce03732b18f759e1d81de
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82043
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>