[!!!][TASK] Merge EXT:t3editor into backend
This patch merges the "t3editor" extension into EXT:backend. The
overall benefit is a globally available syntax highlighting at various
places, without the need to have an extra extension installed.

Another benefit it the removal of installation checks of EXT:t3editor
and fallbacks if the "soft requirement" is not fulfilled, reducing
soft-cross-dependencies.

Resolves: #102440
Releases: main
Change-Id: I76301c9cff08f61a5721fa34f4e3e950687b67b0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81799
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>