[BUGFIX] Restore diff-item-result-inline styling
With the recent introduction of formatted FlexForm diff views, the
styling for "diff-item-result-inline" was removed. However, it is still
needed in other areas. For example in the active TypoScript view when
displaying substitution of constant values.

This patch adds the styling for "diff-item-result-inline" back and in
exchange removes this class from the workspaces "view changes" and
"history" views. There it is not needed, and preserving white-space is
even advantageous.

Resolves: #102413
Related: #91148
Releases: main, 12.4
Change-Id: I6de89ab311a99636771df6eae21ecdadffa09f35
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81803
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>