[TASK] Modify handling of prefixed server version information
Doctrine DBAL deprecated the `ServerInfoAwareConnection`
and `VersionAwarePlatformDriver` interfaces within `3.x`
and force all Driver and DriverConnection to implement
the APIs aware of the server version [1].

TYPO3 used the server version and prefixed them, which
will no longer work with the next Doctrine DBAL major
to due the call chain. The prefixed version has been
used in several places within the reporting module and
backend information to display database vendor aware
information.

This change adds a new `getPlatformServerVersion()`
method to the core connection class and removes the
prefixing from the `getServerVersion()` methods. All
usages in userland code are modified to use the new
method. Additionally, tests are aligned to this change.

[1] https://github.com/doctrine/dbal/blob/3.8.x/UPGRADE.md#deprecated-driver-level-apis-that-dont-take-the-server-version-into-account

Resolves: #102424
Releases: main
Change-Id: I388d6bc79f927179361e8ce4a1486001901c5c38
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81816
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>