[FEATURE] Flush cache tags for file and folder operations
This feature is guarded with toggle "frontend.cache.autoTagging" and
experimental for now: The core flushes cache tags for all kinds of
records automatically when they are created, changed or deleted.
For files and folders, this is not the case.
This feature adds cache tag handling for file or folder operations when
they are created, changed or deleted. Also the file metadata changes are
handled correctly now. This leads to a better
editor experience when the cache tags are used correctly.

Resolves: #102430
Releases: main
Change-Id: I7a624a0f7dbecc4dc97d65e3e2c932060dab06b8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81825
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Garvin Hicking <garvin@hick.ing>