[BUGFIX] Realign additional field options and arguments in scheduler task
When editing a scheduler task, it is possible to select console commands.
These sometimes contain several field options and arguments, displayed as
checkboxes with sometimes additional input fields.

With this patch the position of the checkbox is adjusted to the
FormEngine markup. In addition, additional information about the
individual fields is no longer displayed in the field label itself, but
as a field description below the field label.

The description of the command, which is already part of the selection
field, is no longer displayed below the same field.

Resolves: #102437
Releases: main, 12.4
Change-Id: I2b98f9c7f5c3353a66fa7865b19d9590a96c8265
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81836
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: Jochen Roth <rothjochen@gmail.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Jochen Roth <rothjochen@gmail.com>