[BUGFIX] Suppress command output in functional tests
In #102417 the cli tests were moved to functional
tests. This leads to (unwanted) additional on top of
the dot output in the functional tests.

This has been fixed by using proc_open to catch
the output of the progressbar and friends. Also
the args argument is fixed to allow for multiple
arguments to be used in the data provider.

Resolves: #102453
Releases: main, 12.4
Change-Id: I94ba8a187b5a86aaafeade44b92f9fa88a9b3aba
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81876
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>