[BUGFIX] Set maxlength for varchar inputs in sys_file_metadata
Max values have been set to the input fields in the TCA of the
sys_file_metadata table. The settings have been adjusted to the
set varchar length in the ext_tables.sql file.

Additionally the language column length has been adjusted to
the other fields. The length is set from 12 to 45.

Resolves: #102456
Releases: main, 12.4
Change-Id: Id2b7d510c16b5c2659fa0a7367edeb4f49b782f2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82003
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>