[TASK] Remove `COMPOSER_ALLOW_SUPERUSER=1` from runTests.sh
The environment variable `COMPOSER_ALLOW_SUPERUSER` has been set for
composer command runs to avoid the warning that you are root in the
container.

It has been added to the core-testing-phpXY images so it is now
superflous to set it explicitly on container invocation and can be
removed.

Resolves: #102466
Releases: main, 12.4
Change-Id: Ib91bf883385c524a756b8d2af0fcd80b3d42bc5b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81900
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>