[BUGFIX] Have record icons with human readable title
Add record titles to the record icon alt text. To make
the alt text for record icons helpful, the title and
type of the record is now added to the existing id=x.

To implement this also for the pagetree, the method
BackendUtility::titleAttribForPages() has got a new
parameter to return the value unescaped.
The JsonResponse in the TreeController already
escapes the data and additionally the TemplateResult
from lit also does escaping.

Additionally BackendUtility::getRecordIconAltText() received
a new parameter to make it possible to not escape string values.
This is necessary to prevent values to get double encoded when
used in the fluid context.

Resolves: #102472
Releases: main, 12.4
Change-Id: I2476baccc4caf1ffaf27bbb3d5681cd53aea6052
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81883
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>