[TASK] Cache ordered providers in FormEngine
For large records with many "open" inline children, FormEngine
calculates the same OrderedProviderList over and over again.
This can turn into a bottleneck since DependencyOrderingService
is CPU hungry due to its graph algorithms.

The result of DependencyOrderingService within OrderedProviderList
is a function from given input arguments and can be easily
cached. The patch adds a runtime cache to OrderedProviderList, the
patch can be backported to v13 without side effects.

The patch also reveals some loose ends: It would be much better
to change FormDataGroupInterface to receive the 'providerList'
as argument to compile() which would render implementing services
stateless and shared. This would reduce the number of created
objects in this area to simplify and further improve performance.
It would be even better to turn the lists into service providers
which could be built and cached once at compile time.
This is out of scope of this patch and open for additional patches.

The change gives a measurable performance impact with records
having many "open" inline children: Editing a simple content
elmement with ten attached images reduces the number of expensive
calls to "orderByDependencies()" and saves about 5% server time at
the cost of slightly additional consumed of memmory.

Resolves: #102480
Releases: main, 13.4
Change-Id: Ib6c4865295fa4f669e9a6115753562979ceac5f1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81917
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>