[TASK] Avoid deprecated `Platform->getStringLiteralQuoteCharacter()`
The doctrine team deprecated within `doctrine/dbal 3.x` the
`Platform->getStringLiteralQuoteCharacter()`, postulating
that this is a implementation detail and not a public required
information [1][2].

TYPO3 extends a couple of classes to provide some advanced
abilities to help extension author's avoiding casual safety
traps and additional features. The `ExpressionBuilder` is
one of these classes. Extended methods quotes values per
default, and therefore requires to unquote values at some
points to avoid double quoting of values.

This change removes the call to the deprecated method on
the platform and directly replaces it with the value, as
all supported platforms will return the same quoting char.

Tests are adjusted accordingly.

[1] https://github.com/doctrine/dbal/blob/3.7.x/UPGRADE.md#deprecated-abstractplatform-methods-exposing-quote-characters
[2] https://github.com/doctrine/dbal/pull/5388

Resolves: #102491
Releases: main
Change-Id: I02e0c7dfe5394461f2dded4ec78c4e8e27b0cfd6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81934
Tested-by: Andreas Nedbal <andy@pixelde.su>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Andreas Nedbal <andy@pixelde.su>